+++ src/lib/components/admin/Header/dropdownUsers.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState } from "react";
 import Link from "next/link";
 import Image from "next/image";
 import useLocalStorage from "@lib/hook/useLocalStorage";
-import { User } from "@lib/modal/user";
+import { User } from "@/lib/interface/user.interface";
 import { useSelector } from "react-redux";
 import { RootState } from "@lib/state/store";
