+++ src/lib/interface/category.interface.ts
