+++ src/lib/interface/collection.interface.ts
