+++ src/lib/interface/friend.interface.ts
@@ -1,4 +1,4 @@
-import { User } from "./user";
+import { User } from "./user.interface";
 
 export interface Friend {
     _id?: string;
