+++ src/lib/interface/posts.interface.ts
