+++ src/lib/interface/quizzResult.interface.ts
