+++ src/pages/admin/play/[id].tsx
@@ -3,10 +3,14 @@ import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import Image from "next/image";
-import { getAnwsersIsTrue, getItemQuizz, getQuestionById } from "@/api/quizz";
-import { Anwsers, Question } from "@lib/modal/question";
+import {
+  getAnwsersIsTrue,
+  getItemQuizz,
+  getQuestionById,
+} from "@/api/quizz.api";
+import { Anwsers, Question } from "@/lib/interface/question.interface";
 import { questionInit } from "@lib/config/initQuestion";
-import CountdownTimer from "@lib/components/common/CountdownTimer/DefaultCountdownTimer";
+import CountdownTimer from "@lib/components/common/countdownTimers/defaultCountdownTimers";
 import { addResult, resetResult } from "@lib/state/result/resultSlice";
 import { RootState } from "@lib/state/store";
 import Link from "next/link";
